requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.25
pydantic==2.6.1
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.6